% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_hunt.R
\name{shiny_hunt}
\alias{shiny_hunt}
\title{Calculates number of attempts/hrs to encounter/hatch a shiny}
\usage{
shiny_hunt(
  gen,
  masuda = FALSE,
  shiny_charm = FALSE,
  encounter_rate = 100,
  attempt_time = 15,
  hatch_time = 0
)
}
\arguments{
\item{gen}{Integer denoting generation of pokemon}

\item{masuda}{Is the player using masuda method. Default is FALSE}

\item{shiny_charm}{Does the player have a shiny charm. Default is FALSE}

\item{encounter_rate}{Rate of encounter of the pokemon (only for wild encounters). Default is 100\%}

\item{hatch_time}{Time (in seconds) to hatch a single pokemon egg. Default value is 0}

\item{atempt_time}{Time (in seconds) representing average time taken to encounter a pokemon, or soft reset. Default value is 15 seconds per encounter}
}
\value{
Tibble containing probabilities, number of attempts, and hours
}
\description{
Calculates number of attempts/hrs to encounter/hatch a shiny
}
\examples{
gen = 8
e_rate = 75
att = 20
shiny_hunt(gen = gen, encounter_rate = e_rate, attempt_time = att, shiny_charm = TRUE)
}
