% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boss_completion.R
\name{boss_completion}
\alias{boss_completion}
\title{Title}
\usage{
boss_completion(rates, base_rate = NULL, attempts = NULL, verbose = TRUE)
}
\arguments{
\item{rates}{A list of probabilities as floats between 1 and 0. When base_rate is defined must sum to 1}

\item{base_rate}{A probability between 1 and 0. In the case where there is a fixed rate of receiving an item table roll}

\item{attempts}{Number of attempts. Rounds to nearest int, rounds negatives to 0. Enables function to return probability of completion for number of attempts}

\item{verbose}{Enables printed output}
}
\description{
Title
}
