% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_calc.R
\name{dry_calc}
\alias{dry_calc}
\title{Title}
\usage{
dry_calc(p, n, verbose = TRUE, plot = TRUE)
}
\arguments{
\item{p}{Probability of event occurence.}

\item{n}{The number of attempts.}

\item{verbose}{Controls format of returned probability. Default (True) returns result as a statement, False returns a float.}

\item{plot}{Controls return of plot showing where the resulting probability lies on the binomial distribution. Default is True.}
}
\description{
Title
}
