% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_calc.R
\name{dry_calc}
\alias{dry_calc}
\title{dry_calc}
\usage{
dry_calc(p, n, verbose = TRUE, plot = TRUE)
}
\arguments{
\item{p}{Numeric probability of event occurrence.}

\item{n}{Integer indicating the number of attempts.}

\item{verbose}{Controls format of returned probability; Default (TRUE) prints result as a chr statement; FALSE returns result as a numeric decimal.}

\item{plot}{Controls printing of plot showing where the resulting probability lies on the binomial distribution; Default (TRUE) shows the plot.}
}
\value{
A number, the probability of at least one occurrence of the event given the number of trials is returned as a decimal (if verbose set to FALSE).
}
\description{
Calculates probability of at least one occurrence of an event given the number of attempts.
}
\examples{
# return numeric probability and show plot
dry_calc(0.2, 5L, verbose = FALSE, plot = TRUE)
# return probability in a chr statement and don't show plot
dry_calc(0.5, 3L, verbose = TRUE, plot = FALSE)
}
