% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_hunt.R
\name{shiny_hunt}
\alias{shiny_hunt}
\title{Calculates number of attempts/hrs to encounter/hatch a shiny}
\usage{
shiny_hunt(
  gen,
  masuda = FALSE,
  shiny_charm = FALSE,
  encounter_rate = 100,
  attempt_time = 15,
  hatch_time = 0
)
}
\arguments{
\item{gen}{Integer denoting generation of pokemon}

\item{masuda}{Is the player using masuda method}

\item{shiny_charm}{Does the player have a shiny charm}

\item{encounter_rate}{Rate of encounter of the pokemon (only for wild encounters)}

\item{hatch_time}{Time (in seconds) to hatch a single pokemon egg}

\item{atempt_time}{Time (in seconds) representing average time taken to encounter a pokemon, or soft reset}

\item{verbose}{Controls format of returned probability. Default (True) prints results as statements, False returns a dict.}
}
\value{
Dictionary containing probabilities as keys and number of attempts/hrs values as tuples
}
\description{
Calculates number of attempts/hrs to encounter/hatch a shiny
}
